#14-change-group-command: rename `ScheduleCommand` to `UserCommand`
