+++ src/main/java/org/anstreth/schedulebot/commands/UserCommand.java
@@ -1,5 +1,5 @@
 package org.anstreth.schedulebot.commands;
 
-public enum ScheduleCommand {
+public enum UserCommand {
     UNKNOWN, TODAY, TOMORROW, WEEK
 }
